[build] Improve torque explicit dependencies
Add a couple of missing dependencies to the build rules for
torque-generated files. Also add a allow_circular_includes_from
declaration to v8_initializers, so that its files can be accessed by
torque_generated_initializers which cannot depend on v8_initializers
without introducing a cycle.

This reduces the number of gn check errors with --check-generated from
10512 to 33.

Bug: v8:7330
Change-Id: I324295c639094c258b3e8e674aef6edc043fe505