[torque] Don't report unused macros if they're marked as unused
Torque generally supports marking things as unused by adding a single
underscore to the front of identifiers. For locals, this feature works
fine. For macros, it's only half-implemented: attempting to look up a
macro that starts with _ fails (as it should), but the compiler also
complains if the macro is unused. This change avoids emitting the latter
error.

Bug: v8:7793
Change-Id: Ib021c053004a180dd31993d3ad06cde463bafd5d