Collect receiver to feedback for prototype.apply and prototype.call
When a function is invoked by prototype.call or prototype.apply, it may undergo following transformation in the JSCallReducer:
	receiver.apply(this, args) ->
	this.receiver(...args)
Since the new target (also the receiver of call()/apply()) is not collected to the feedback slot, further speculative optimization on the new target is not available if the new target is not a heapconstant.

With this CL, the receiver will be collected to the feedback instead of the target if the target is a prototype.call or prototype.apply. It may improve the performance of the following usecase by ~80%.

function reduceArray(func, arr, r) {
    for (var i = 0, len = arr.length; i < len; i++) {
            r = func.call(null, r, arr[i]);
    }
    return r;
}

var a = 0;
for (var i = 0; i < 10000000; i++) {
    a += reduceArray(Math.imul, [5,6,2,3,7,6,8,3,7,9,2,5,], 1);
}
console.log(a);

This CL also improves the runTime score of JetStream2/richards-wasm by ~45% in default, ~60% with --turbo-inline-js-wasm-calls.

Change-Id: I542eb8d3fcb592f4e0993af93ba1af70e89c3982