Merged: [wasm-simd][liftoff][x64][ia32] Fix i64x2.mul codegen
We are overwriting rhs when dst == rhs && dst != lhs. This is not a
problem on TurboFan because we specify unique registers and dst == lhs
in the instruction-selector.

The fix is to use the helper EmitSimdCommutativeBinOp, which will check
for dst == rhs (pmuludq is commutative).

(cherry picked from commit 7514db4b605d9a003c862fd649a8a163105d9f67)

Tbr: gdeepti@chromium.org
Bug: v8:11612
Change-Id: I833a6bb21d41ca173ee36353758682128f706b8b