[Liftoff] Unify builtin calling logic
This factors out the duplicated logic for calling builtins in Liftoff.
Recording a source position is optional, but many call sites so far
missed it even though it would be helpful for debugging e.g. in case of
OOM errors. Thus this CL adds source positions for most builtin calls.

R=ahaas@chromium.org

Change-Id: Id4e3e2d0177dbb5dded533ed3ec294bdbec95e9f