[ptr-cage] Factor CodeRange out of MemoryAllocator
This CL factors out a CodeRange class out of MemoryAllocator for use in
the shared pointer cage. When V8_COMPRESS_POINTERS_IN_SHARED_CAGE is
defined, there is a process-wide CodeRange that is shared by all
Isolates in the process.

Since pointer cage is still experimental, to avoid API churn this
CodeRange's size is not configurable and is always the maximal size
depending on the underlying platform.

Change-Id: Ie94f52746f2c5450247a999cc6071e3914d4cf0c