[ptr-cage] Factor out a VirtualMemoryCage and remove PtrComprCage
The plan is to use VirtualMemoryCage both for the pointer compression
cage as well as the code range in a future CL. The PtrComprCage class is
removed in favor of using VirtualMemoryCage directly.

This CL also simplifies the pointer compression reservation strategy.
The existing, more complex reservation strategy is a holdover from the
now abandoned "base in the middle" layout that was not simple alignment.

Bug: v8:11460
Change-Id: I4e34a3db1359319e3539ede587f6a73e0af03eec