Fix for issue 1156498. Crash on  break on throw in static initializer.
Fixed an issue where the "Pause on exceptions" feature of the debugger
failed  a DCHECK if the throw happened on the declaration of a static
variable inside a class inside a function that needs context.

The DCHECK_IMPLES inside the ScopeIterator::Type() FUNCTION SCOPE case
failed because the function needs context, but the context
scope is set to the class scope, which is the top element of the scope
chain.

This change adds a new conditional to turn on the ignore_nested_scopes
flag in the mentioned scenario. This triggers the context to
iterate over its closures until it finds one with a function scope.

The change causes an intended change on how the chain scope is inspected
when a class is defined inside a function, hence the expected output
of the inspector test is also modified for this change.



Bug: chromium:1156498
Change-Id: I2ef7dbe5a251b57c52fde7994086247438c669e3