[compiler] Do Map::inobject_properties_start non-atomically for no-cm
In https://chromium-review.googlesource.com/c/v8/v8/+/2843359 I
changed the kInObjectPropertiesStartOrConstructorFunctionIndexOffset
to be set non-atomically for both concurrent marking and without.
According to this flaky failure
https://logs.chromium.org/logs/v8/buildbucket/cr-buildbucket.appspot.com/8848894352656677568/+/u/Check__flakes_/function-without-prototype
the concurrent marker can see maps even when they are still in
factory.cc.

We should be okay to perform the set/get non-atomically when
concurrent marking is off.

Bug: v8:7790
Change-Id: I6e6818af9d18ced1c76022bcc0e74539ef0bc483