[turbofan] Remove control input from Projection
The control input was unused. We would like to remove it as it might
create problems in the future in the following scenario: If multiple
projected nodes get merged during e.g. value numbering reduction, we end
up with multiple projections of the same node (with the same index).
Those will not be able to be merged if they have different control
inputs, resulting in an invalid graph. This has been avoided so far, at
least partly by using graph()->start() as the control input for
projections, but it might break in the future.

Change-Id: I22729123a1c8dec56e1adabc95346c317e51cde7