[sandbox][x64] ArrayBuffer caging prototype, variant 5
In this variant, the data pointers for ArrayBuffers, TypedArrays, and
DataViews are stored as shifted pointers with the coresponding base
address being a constant.

Due to the cage, nullptr is now generally an invalid backing store
pointer. As such, most places have explicit nullptr handling added so
they do not "uncage" nullptr. The only exception is in turbofan (in
memory lowering), where these checks can (apparently) be omitted as it
is never valid to load nullptr there.

This change emulates a cage with base address 0xffffffff00000000 and
size 0x1000000000000 so as to not have to modify the ArrayBuffer
allocator. A base != zero is required to stop turbofan from optimizing
away parts of the uncaging procedure. This should still allow accurate
performance measurements of the impact of such a cage on V8 performance
as the engine will perform the same operations as if the buffers were
actually caged.

This cage is generally incompatible with on-heap TypedArrays. Those
likely either have to be disabled (currently the case) or be implemented
with an additional branch on every access.

Commit: false
Bug: v8:10391
Change-Id: I9309825eb623e14f942389887d77dfb7a79b6aa4