[builtins][x64] Use callee-saved registers for write barrier stubs
Calls to the record write stub are quite frequent and the caller has to
save all registers used by the builtin.

This CL moves the register saving to the builtin itself, reducing the
call-site code size significantly in many cases.

CallRecordWriteStubSaveRegisters preserves the existing behaviour and
saves clobbered registers.
CallRecordWriteStub expects the registers to match the ones specified
in the WriteBarrierDescriptor for more compact code.

Follow-up CLs with introduce the same behaviour to other platforms.

Bug: v8:11420
Change-Id: I3a0c500fbe26f82ee2243a61dbf574fd31656982